+++ odoo/custom_addons/pim_ext/models/product.py
@@ -9,7 +9,6 @@ import pytz
 
 from lxml import etree
 from odoo import models, api, fields,_, tools
-# from odoo.addons.test_impex.models import field
 from odoo.exceptions import UserError, ValidationError
 import traceback,pdb,inspect
 
@@ -19,853 +18,6 @@ from googletrans import Translator
 
 
 _logger = logging.getLogger(__name__)
-
-
-
-class AttributeForm(models.Model):
-     _inherit = 'product.attribute'
-     _rec_name = 'name'
-
-     display_type = fields.Selection(
-          selection=[
-               ('radio', 'Radio'),
-               ('date', 'Date'),
-               ('file', 'File'),
-               ('identifier', 'Identifier'),
-               ('image', 'Image'),
-               # ('measurement', 'Measurement'),
-               ('multi_select', 'Multi Select'),
-               ('multi_checkbox', 'Multi Checkbox'),
-               ('link', 'Link'),
-               ('number', 'Integer'),
-               ('price', 'Price'),
-               ('ref_data_multi', 'Reference Data Multi Select'),
-               ('ref_data_simple_select', 'Reference Data Simple Select'),
-               ('simple_select', 'Simple Select'),
-               ('text', 'Text'),
-               ('textarea', 'Text Area'),
-               ('yes_no', 'Checkbox'),
-               ('pills', 'Pills'),
-               ('select', 'Select'),
-               ('color', 'Color'),
-               ('multi', 'Multi-checkbox (option)'),
-          ],
-          required=True,
-          help="The display type used in the Product Configurator.")
-     attribute_group = fields.Many2one('attribute.group', string='Attribute Group', tracking=True)
-     attribute_type_id = fields.Many2one('pim.attribute.type', string='PIM Attribute Type')
-     is_mandatory = fields.Boolean(string='Mandatory', default=False)
-     is_required_in_clone = fields.Boolean(string='Required in Clone', default=True)
-     is_cloning = fields.Boolean(string='Cloning', default=True)
-     is_completeness = fields.Boolean(string='Completeness')
-     original_name = fields.Char('Previous Name')
-
-     attribute_types = fields.Selection([('basic', 'Basic'),
-                                         ('optional', 'Optional')], string='Attribute Type')
-     attribute_types_id = fields.Many2one('pim.attribute.type', string='Attribute Type')
-     completed_in_percent = fields.Float('Completed Progressbar',compute="_compute_completness")
-     state = fields.Selection([('unpublish', 'Unpublish'), ('publish', 'Publish')], string='Status', default='unpublish')
-
-
-     position_ref_field_id = fields.Many2one("ir.model.fields", string="Position After", domain="[('model_id.model','=','product.management')]")
-
-     code = fields.Char(string='Code')
-
-     unique_value = fields.Selection(
-          [('yes', 'Yes'), ('no', 'No')],
-          string='Unique Value',
-     )
-
-     value_per_channel = fields.Selection(
-          [('yes', 'Yes'), ('no', 'No')],
-          string='Value per channel',
-     )
-
-     value_per_locale = fields.Selection(
-          [('yes', 'Yes'), ('no', 'No')],
-          string='Value per locale',
-     )
-
-     usable_in_grid = fields.Selection(
-          [('yes', 'Yes'), ('no', 'No')],
-          string='Usable in grid',
-     )
-
-     locale_specific = fields.Selection(
-          [('yes', 'Yes'), ('no', 'No')],
-          string='Locale specific',
-     )
-
-     master_attribute_ids = fields.One2many(
-          'product.attribute',
-          'parent_id',
-          compute='_compute_master_attribute_ids',
-          string='Attributes'
-     )
-
-     parent_id = fields.Many2one(
-          'product.attribute',
-          string='Parent Group',
-          ondelete='cascade',
-     )
-
-     label_transaltion = fields.Char(string='English', compute="_compute_label_attribute_translation")
-
-     user_id = fields.Many2one('res.users', string="User", default=lambda self: self.env.user)
-
-     history_log = fields.Html(string='History Log', help="This field stores the history of changes.")
-
-     products_id = fields.Many2one('product.template', string='Products')
-
-     @api.depends('name')
-     def _compute_label_attribute_translation(self):
-          translator = Translator()
-          for record in self:
-               if record.name:
-                    try:
-                         user_lang = self.env.user.lang
-                         lang_rec = self.env['res.lang'].search([('code', '=', user_lang)], limit=1)
-                         src_lang = lang_rec.url_code
-                         translation = translator.translate(record.name, src=src_lang, dest='en')
-                         record.label_transaltion = translation.text.capitalize()
-                    except Exception as e:
-                         record.label_transaltion = 'Error in translation'
-               else:
-                    record.label_transaltion = ''
-
-     # this is for loading tree view in while click edit button
-     def _compute_master_attribute_ids(self):
-          for record in self:
-               record.master_attribute_ids = self.env['product.attribute'].search([])
-
-     def save_attributes(self):
-          return {
-               'type': 'ir.actions.act_window',
-               'view_mode': 'form',
-               'res_model': 'pim.attribute.type',
-               'view_id': self.env.ref('pim_ext.view_product_attribute_custom').id,
-               'res_id': self.id,
-               'context': {'no_breadcrumbs': True},
-               'target': 'current',
-          }
-
-     # attribute master edit button
-     def attribute_edit_open_form_view(self):
-          self.ensure_one()
-          return {
-               'type': 'ir.actions.act_window',
-               'name': 'Edit Attribute',
-               # 'res_model': 'product.attribute',
-               'view_mode': 'form',
-               'res_model': 'pim.attribute.type',
-               'view_id': self.env.ref('pim_ext.view_product_attribute_custom').id,
-               'context': {'no_breadcrumbs': True,
-                           'default_attribute_ids': self.env['product.attribute'].search([]).ids,},
-               'res_id': self.attribute_type_id,
-          }
-
-     # delete button attribute
-     def attribute_master_unlink(self):
-          return {
-               'name': 'Confirm Deletion',
-               'type': 'ir.actions.act_window',
-               'res_model': 'attribute.master.unlink.wizard',
-               'view_mode': 'form',
-               'target': 'new',
-               'context': {'default_attribute_id': self.id},
-          }
-
-     def action_back_to_attribute_menu(self):
-          menu_id = self.env.ref('pim_ext.menu_pim_attribute_action')
-          return {
-               'type': 'ir.actions.client',
-               'tag': 'reload',
-               'params': {
-                    'menu_id': menu_id.id,
-               },
-          }
-
-
-     def action_publish_attribute(self):
-          for record in self:
-               if not record.attribute_group:
-                    raise ValidationError("Alert!! Please select the group of the '%s' attribute." % record.name)
-               if record.state == 'unpublish':
-                    if record.name:
-                         field_name = f"x_{record.name.strip().replace(' ', '_').lower()}"
-                         field_type = None
-                         field_data = {}
-                         if record.display_type == 'text':
-                              field_type = 'char'
-                         elif record.display_type == 'textarea':
-                              field_type = 'text'
-                         elif record.display_type == 'number':
-                              field_type = 'integer'
-                         elif record.display_type == 'file':
-                              field_type = 'binary'
-                         elif record.display_type == 'image':
-                              field_type = 'binary'
-                         elif record.display_type == 'simple_select':
-                              field_type = 'selection'
-                         elif record.display_type == 'date':
-                              field_type = 'date'
-                         if not field_type:
-                              continue
-
-                         field_data.update({
-                              'name': field_name,
-                              'field_description': record.name,
-                              'ttype': field_type,
-                              'model_id': self.env.ref('pim_ext.model_product_management').id,
-                              'state': 'manual',
-                         })
-                         create_field = self.env['ir.model.fields'].create(field_data)
-                         if field_type == 'selection' and record.value_ids:
-                              for value in record.value_ids:
-                                   sel_value = value.name.strip().replace(' ', '_').lower()
-                                   sel_name = value.name
-                                   self.env['ir.model.fields.selection'].create({
-                                        'name': sel_name,
-                                        'value': sel_value,
-                                        'field_id': create_field.id
-                                   })
-                         view = self.env.ref('pim_ext.product_managemnt_form_view')
-                         widget = "image" if record.display_type == "image" else ""
-                         widget_attribute = f' widget="{widget}"' if widget else ""
-                         new_field_xml = f'<field name="{field_name}"{widget_attribute}  required="False"/>'
-                         arch_value = f"""
-                         <xpath expr="//field[@name='{self.position_ref_field_id.name}']" position="after">
-                             {new_field_xml}
-                         </xpath>
-                         """
-                         self.env['ir.ui.view'].sudo().create({
-                              'name': f'add_field_{field_name}_to_product_management',
-                              'type': 'form',
-                              'model': view.model,
-                              'inherit_id': view.id,
-                              'arch': arch_value,
-                         })
-                         group_line = self.env['attribute.group.lines'].search([
-                              ('attr_group_id', '=', record.attribute_group.id),
-                              ('product_attribute_id', '=', record.id)
-                         ], limit=1)
-
-                         if not group_line:
-                              self.env['attribute.group.lines'].create({
-                                   'attr_group_id': record.attribute_group.id,
-                                   'product_attribute_id': record.id,
-                                   'display_type': record.display_type,
-                              })
-
-                         record.write({'state': 'publish'})
-
-
-     # def action_update_publish_attribute(self):
-     #      for record in self:
-     #           if not record.original_name:
-     #                record.original_name = record.name
-     #                print(record.original_name, 'originalllllllllllllll')
-     #
-     #           if record.name != record.original_name:
-     #                field_name = f"x_{record.original_name.strip().replace(' ', '_').lower()}"
-     #                ir_field = self.env['ir.model.fields'].sudo().search([
-     #                     ('name', '=', field_name),
-     #                     ('model_id', '=', self.env.ref('pim_ext.model_product_management').id)
-     #                ], limit=1)
-     #                if ir_field:
-     #                     ir_field.write({
-     #                          'field_description': record.name,
-     #                     })
-     #                     new_field_name = f"x_{record.name.strip().replace(' ', '_').lower()}"
-     #                     self.env.cr.execute(f"""
-     #                     ALTER TABLE product_management RENAME COLUMN {field_name} TO {new_field_name}
-     #                 """)
-     #                record.original_name = record.name
-     #      return {
-     #           'type': 'ir.actions.client',
-     #           'tag': 'display_notification',
-     #           'params': {
-     #                'message': 'Attributes updated successfully, including related fields!',
-     #                'type': 'success',
-     #                'sticky': False,
-     #           },
-     #      }
-
-     def _compute_completness(self):
-          for rec in self:
-               data = rec.search_read([('id','=',rec._origin.id)],fields=['attribute_type_id','display_type','attribute_types_id','attribute_types','attribute_group','is_mandatory','is_required_in_clone', 'is_cloning'])
-               false_count = sum(1 for d in data for value in d.values() if value != False)
-               rec.completed_in_percent = (false_count/7) * 100
-               # raise ValidationError(false_count/8)
-
-     def create_pim_attribute_type(self):
-          # Return the action to open the product.attribute custom layout
-          return {
-               'type': 'ir.actions.act_window',
-               'name': 'Create Attribute',
-               'res_model': 'pim.attribute.type',
-               'view_mode': 'form',
-               'view_id': self.env.ref('pim_ext.view_product_attribute_custom').id,
-               'target': 'current',
-               'context': {
-                    'default_attribute_ids': self.env['product.attribute'].search([]).ids,
-               },
-          }
-
-     @api.constrains('name', 'value_ids')
-     def check_attribute_name(self):
-          pattern = "^(?=.*[a-zA-Z0-9])[A-Za-z0-9 ]+$"
-          for rec in self:
-               name_count = self.search_count([('name', '=ilike', rec.name)])
-               if name_count > 1:
-                    raise ValidationError("Attribute name already exist")
-               if rec.name and not re.match(pattern, rec.name):
-                    raise ValidationError("Attribute Name should be AlphaNumeric")
-
-     def _log_changes(self, rec, vals, action):
-          # Get the current time in UTC
-          updated_write_date_utc = fields.Datetime.now()
-          # Convert to the user's timezone
-          user_tz = self.env.user.tz or 'UTC'  # Default to UTC if no timezone is set
-          updated_write_date = updated_write_date_utc.astimezone(pytz.timezone(user_tz)).strftime("%d/%m/%Y %H:%M:%S")
-          new_write_uid = self.env.user.display_name
-
-          changes = []  # Store changes in list
-
-          for key in vals:
-               if key in ['name', 'code', 'active', 'create_variant', 'sequence', 'attribute_group',
-                          'display_type', 'attribute_type_id', 'is_mandatory', 'is_required_in_clone',
-                          'is_cloning', 'is_completeness', 'original_name', 'attribute_types',
-                          'attribute_types_id', 'completed_in_percent', 'state', 'position_ref_field_id',
-                          'unique_value', 'value_per_channel', 'value_per_locale', 'usable_in_grid',
-                          'locale_specific', 'master_attribute_ids', 'label_transaltion']:
-
-                    attribute = rec._fields[key].string
-
-                    if key == 'attribute_group':
-                         old_value = rec.attribute_group.display_name if rec.attribute_group else 'N/A'
-                         new_value = self.env['attribute.group'].browse(vals[key]).display_name if vals.get(
-                              key) else 'N/A'
-                    else:
-                         # Set old_value to 'N/A' if creating a new record
-                         old_value = getattr(rec, key, 'N/A') if action == "update" else 'N/A'
-                         new_value = vals[key] or 'N/A'
-
-                    # Formatting Old & New values on the same line with space
-                    change_entry = f"""
-                  <li>
-                      <strong>{attribute}</strong><br>
-                      <span style='color: red;'>Old value:</span> {old_value}  
-                      &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                      <span style='color: green;'>New value:</span> {new_value}
-                  </li>
-                  """
-                    changes.append(change_entry)
-
-          if changes:
-               action_text = "Created by" if action == "create" else "Updated by"
-               user_info = f"<small>{action_text} <strong>{new_write_uid}</strong> on {updated_write_date}</small>"
-
-               full_message = f"""
-              <div style="border-left: 3px solid #6C757D; padding-left: 10px; margin-bottom: 15px;">
-                  {user_info}
-                  <ul style="list-style-type: none; padding-left: 0;">{''.join(changes)}</ul>
-              </div>
-              """
-
-               rec.history_log = tools.html_sanitize(full_message) + (rec.history_log or '')
-
-     def create(self, vals):
-          # Create the record
-          record = super(AttributeForm, self).create(vals)
-          self._log_changes(record, vals, action="create")
-          return record
-
-     def write(self, vals):
-          for rec in self:
-               self._log_changes(rec, vals, action="update")
-          return super(AttributeForm, self).write(vals)
-
-
-class AttributeMasterUnlinkWizard(models.TransientModel):
-    _name = 'attribute.master.unlink.wizard'
-    _description = 'Wizard to Confirm Deletion of Attribute master'
-
-    attribute_id = fields.Many2one('product.attribute', string="Attribute Master")
-
-    def confirm_unlink(self):
-        menu_id = self.env['ir.ui.menu'].search([('name', '=', 'Attributes')])
-        if self.attribute_id:
-             self.attribute_id.unlink()
-        return {
-             'type': 'ir.actions.client',
-             'tag': 'reload',
-             'params': {
-                  'menu_id': menu_id.id,
-             },
-        }
-
-class AttributeGroup(models.Model):
-     _name = 'attribute.group'
-     _inherit = ['mail.thread', 'mail.activity.mixin']
-
-     name = fields.Char('Name', required=True,)
-     description = fields.Text(string='Description')
-     active = fields.Boolean('Active', default=True)
-     attributes_ids = fields.One2many('product.attribute', 'attribute_group', string='Attributes', required=True, tracking=True, store=True)
-     attribute_family_id = fields.Many2many('family.attribute', string='Attribute Family', required=True, tracking=True, store=True)
-     attribute_group_line_ids = fields.One2many('attribute.group.lines', 'attr_group_id', string='Attribute group line')
-     attribute_code = fields.Selection([('medias', 'Medias')], string="Code")
-     attribute_code_rec = fields.Char(string='Code', readonly=True)
-     attribute_label = fields.Char(string='English', compute="_compute_label_translation")
-     parent_id = fields.Many2one(
-          'attribute.group',
-          string='Parent Group',
-          ondelete='cascade',
-     )
-     child_ids = fields.One2many(
-          'attribute.group',
-          'parent_id',
-          compute='_compute_child_ids',
-          string='Sub-groups'
-     )
-     user_id = fields.Many2one('res.users', string="User", default=lambda self: self.env.user)
-
-     history_log = fields.Html(string='History Log', help="This field stores the history of changes.")
-
-     is_create_mode = fields.Boolean(default=False, string='Create Mode')
-
-     def _log_changes(self, rec, vals, action):
-          updated_write_date_utc = fields.Datetime.now()
-          user_tz = self.env.user.tz or 'UTC'
-          updated_write_date = updated_write_date_utc.astimezone(pytz.timezone(user_tz)).strftime("%d/%m/%Y %H:%M:%S")
-          new_write_uid = self.env.user.display_name
-
-          changes = []
-
-          # Fields to track in attribute.group
-          tracked_fields = ['name', 'attribute_code_rec', 'active', 'parent_id', 'description', 'attribute_code']
-
-          # Fields to track in attribute.group.lines
-          tracked_group_line_fields = [
-               'product_attribute_id', 'display_type', 'enable', 'value_per_channel', 'value_per_locale'
-          ]
-
-          # Track changes in attribute.group fields
-          for key in vals:
-               if key in tracked_fields:
-                    attribute = rec._fields[key].string
-                    old_value = getattr(rec, key, 'N/A') if action == "update" else 'N/A'
-                    new_value = vals[key] or 'N/A'
-
-                    if key == 'parent_id':
-                         old_value = rec.parent_id.display_name if rec.parent_id else 'N/A'
-                         new_value = self.env['attribute.group'].browse(vals[key]).display_name if vals.get(
-                              key) else 'N/A'
-
-                    change_entry = f"""
-                  <li>
-                      <strong>{attribute}</strong><br>
-                      <span style='color: red;'>Old value:</span> {old_value}  
-                      &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                      <span style='color: green;'>New value:</span> {new_value}
-                  </li>
-                  """
-                    changes.append(change_entry)
-
-          # Track changes in attribute_group_line_ids (One2many)
-          if 'attribute_group_line_ids' in vals:
-               for command in vals['attribute_group_line_ids']:
-                    if command[0] == 1:  # Update existing record
-                         line_id = self.env['attribute.group.lines'].browse(command[1])
-                         line_changes = []
-                         for field in tracked_group_line_fields:
-                              if field in command[2]:
-                                   old_value = getattr(line_id, field, 'N/A')
-                                   new_value = command[2][field] or 'N/A'
-
-                                   if field == 'product_attribute_id':
-                                        old_value = line_id.product_attribute_id.display_name if line_id.product_attribute_id else 'N/A'
-                                        new_value = self.env['product.attribute'].browse(
-                                             new_value).display_name if new_value else 'N/A'
-
-                                   line_changes.append(f"""
-                              <li>
-                                  <strong>{line_id._fields[field].string}</strong><br>
-                                  <span style='color: red;'>Old:</span> {old_value}  
-                                  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                                  <span style='color: green;'>New:</span> {new_value}
-                              </li>
-                              """)
-
-                         if line_changes:
-                              changes.append(f"""
-                          <li>
-                              <strong>Updated Attribute Group Line (ID {line_id.id})</strong>
-                              <ul>{''.join(line_changes)}</ul>
-                          </li>
-                          """)
-
-
-
-                    elif command[0] == 0:  # New record
-
-                         new_values = command[2]  # Get the dictionary of new values
-
-                         new_line_changes = []  # Store field changes
-
-                         for field in tracked_group_line_fields:
-
-                              if field in new_values:
-
-                                   field_label = self.env['attribute.group.lines']._fields[field].string
-
-                                   new_value = new_values[field] or 'N/A'
-
-                                   old_value = "N/A"  # Since it's a new record, old value is always N/A
-
-                                   # Fetch display_name for Many2one fields
-
-                                   if field == 'product_attribute_id':
-                                        old_value = "N/A"  # New record, so old value is not present
-
-                                        new_value = self.env['product.attribute'].browse(
-                                             new_value).display_name if new_value else 'N/A'
-
-                                   new_line_changes.append(f"""
-
-                                <li>
-
-                                    <strong>{field_label}</strong><br>
-
-                                    <span style='color: red;'>Old value:</span> {old_value}  
-
-                                    &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-
-                                    <span style='color: green;'>New value:</span> {new_value}
-
-                                </li>
-
-                                """)
-
-                         if new_line_changes:
-                              change_entry = f"""
-
-                            <li>
-
-                                <strong>New Attribute Group Line Added:</strong><br>
-
-                                <ul>{''.join(new_line_changes)}</ul>
-
-                            </li>
-
-                            """
-
-                              changes.append(change_entry)
-
-
-
-
-                    elif command[0] == 2:  # Deletion (Removed record)
-
-                         line_id = self.env['attribute.group.lines'].browse(command[1])
-
-                         removed_line_changes = []  # Store field details of removed line
-
-                         for field in tracked_group_line_fields:
-
-                              field_label = line_id._fields[field].string
-
-                              old_value = getattr(line_id, field, 'N/A')
-
-                              new_value = "N/A"  # Since it's a deletion, new value is always N/A
-
-                              # Fetch display_name for Many2one fields
-
-                              if field == 'product_attribute_id':
-                                   old_value = line_id.product_attribute_id.display_name if line_id.product_attribute_id else 'N/A'
-
-                              removed_line_changes.append(f"""
-
-                            <li>
-
-                                <strong>{field_label}</strong><br>
-
-                                <span style='color: red;'>Old value:</span> {old_value}  
-
-                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-
-                                <span style='color: green;'>New value:</span> {new_value}
-
-                            </li>
-
-                            """)
-
-                         if removed_line_changes:
-                              change_entry = f"""
-
-                            <li>
-
-                                <strong>Removed Attribute Group Line: </strong><br>
-
-                                <ul>{''.join(removed_line_changes)}</ul>
-
-                            </li>
-
-                            """
-
-                              changes.append(change_entry)
-
-          if changes:
-               action_text = "Created by" if action == "create" else "Updated by"
-               user_info = f"<small>{action_text} <strong>{new_write_uid}</strong> on {updated_write_date}</small>"
-               full_message = f"""
-              <div style="border-left: 3px solid #6C757D; padding-left: 10px; margin-bottom: 15px;">
-                  {user_info}
-                  <ul style="list-style-type: none; padding-left: 0;">{''.join(changes)}</ul>
-              </div>
-              """
-               rec.history_log = tools.html_sanitize(full_message) + (rec.history_log or '')
-
-     def write(self, vals):
-          for rec in self:
-               self._log_changes(rec, vals, action="update")
-          return super(AttributeGroup, self).write(vals)
-
-     def create(self, vals):
-          vals['attribute_code_rec'] = self.env['ir.sequence'].next_by_code(
-               'attribute.group') or None
-          record = super(AttributeGroup, self).create(vals)
-          self._log_changes(record, vals, action="create")
-
-          return record
-
-     @api.depends('name')
-     def _compute_label_translation(self):
-          translator = Translator()
-          for record in self:
-               if record.name:
-                    try:
-                         user_lang = self.env.user.lang
-                         lang_rec = self.env['res.lang'].search([('code', '=', user_lang)], limit=1)
-                         src_lang = lang_rec.url_code
-                         translation = translator.translate(record.name, src=src_lang, dest='en')
-                         record.attribute_label = translation.text.capitalize()
-                    except Exception as e:
-                         record.attribute_label = 'Error in translation'
-               else:
-                    record.attribute_label = ''
-
-     # this is for loading tree view in while click edit button
-     def _compute_child_ids(self):
-          for record in self:
-               record.child_ids = self.env['attribute.group'].search([])
-
-     # attribute group edit button
-     def attribute_group_open_form_view(self):
-          all_attribute_group_ids = self.env['attribute.group'].search([]).ids
-          self.ensure_one()
-          return {
-               'type': 'ir.actions.act_window',
-               'name': 'Edit Attribute Group',
-               'res_model': 'attribute.group',
-               'view_mode': 'form',
-               'view_id': self.env.ref('pim_ext.view_product_attribute_groups_custom').id,
-               'context': {'no_breadcrumbs': True,
-                           'default_is_create_mode': False,
-                           },
-               'res_id': self.id,
-          }
-
-     # delete button group
-     def attribute_group_unlink(self):
-          return {
-               'name': 'Confirm Deletion',
-               'type': 'ir.actions.act_window',
-               'res_model': 'attribute.group.unlink.wizard',
-               'view_mode': 'form',
-               'target': 'new',
-               'context': {'default_group_id': self.id},
-          }
-
-
-     def action_back_to_menu(self):
-          menu_id = self.env.ref('pim_ext.menu_pim_attribute_action2')
-
-          # return {
-          #      'type': 'ir.actions.act_window',
-          #      'res_model': 'attribute.group',
-          #      'view_mode': 'list',
-          #      'view_id': action_id,
-          #      # 'context': {'no_breadcrumbs': True},
-          #      'target': 'current',
-          # }
-          return {
-               'type': 'ir.actions.client',
-               'tag': 'reload',
-               'params': {
-                    'menu_id': menu_id.id,
-               },
-          }
-
-     def save_attributes(self):
-          for line in self.attribute_group_line_ids:
-               if not line.product_attribute_id:
-                    raise UserError(f"No product attribute linked in line {line.id}.")
-
-               # Update the product.attribute if needed
-               line.product_attribute_id.write({
-                    'attribute_group': self.id,
-               })
-          return {
-               'type': 'ir.actions.act_window',
-               'view_mode': 'form',
-               'res_model': 'attribute.group',
-               'view_id': self.env.ref('pim_ext.view_product_attribute_groups_custom').id,
-               'res_id': self.id,
-               'context': {'no_breadcrumbs': True},
-               'target': 'current',
-          }
-
-     def create_pim_attribute_groups(self):
-
-          return {
-               'type': 'ir.actions.act_window',
-               'name': 'Create Attribute group',
-               'res_model': 'attribute.group',
-               'view_mode': 'form',
-               'view_id': self.env.ref('pim_ext.view_product_attribute_groups_custom').id,
-               'target': 'current',
-               'context': {
-                    'default_is_create_mode': True,
-                    'default_child_ids': self.env['attribute.group'].search([]).ids,
-               },
-          }
-
-     # @api.model
-     # def create(self, vals):
-     #      print('fkdjfkdfd', vals)
-     #      record = super(Attributegroup, self).create(vals)
-     #      record._check_unique_attributes()
-     #      return record
-     #
-     # @api.model
-     # def write(self, vals):
-     #      print('fko9444444444', vals)
-     #      res = super(Attributegroup, self).write(vals)
-     #      if 'attribute_group_line_ids' in vals:
-     #           self._check_unique_attributes()
-     #      return res
-     #
-     # @api.constrains('attribute_group_line_ids')
-     # def _check_unique_attributes(self):
-     #      for record in self:
-     #           attributes_in_group = record.attribute_group_line_ids.mapped('product_attribute_id')
-     #           for attribute in attributes_in_group:
-     #                other_groups = self.env['attribute.group'].search([
-     #                     ('id', '!=', record.id),
-     #                     ('attribute_group_line_ids.product_attribute_id', '=', attribute.id)
-     #                ])
-     #                if other_groups:
-     #                     raise ValidationError(
-     #                          f"The attribute '{attribute.name}' is already assigned to another attribute group."
-     #                     )
-
-
-class AttributeGroupLine(models.Model):
-     _name = 'attribute.group.lines'
-
-     attr_group_id = fields.Many2one('attribute.group', string='Attribute Group')
-     used_attribute_ids = fields.Many2many(
-          'product.attribute',
-          compute='_compute_used_attribute_ids',
-          string='Used Attributes',
-          store=False
-     )
-     product_attribute_id = fields.Many2one('product.attribute', string='Product Attribute')
-     display_type = fields.Selection(
-          selection=[
-               ('radio', 'Radio'),
-               ('date', 'Date'),
-               ('file', 'File'),
-               ('identifier', 'Identifier'),
-               ('image', 'Image'),
-               ('measurement', 'Measurement'),
-               ('multi_select', 'Multi Select'),
-               ('link', 'Link'),
-               ('number', 'Integer'),
-               ('price', 'Price'),
-               ('ref_data_multi', 'Reference Data Multi Select'),
-               ('ref_data_simple_select', 'Reference Data Simple Select'),
-               ('simple_select', 'Simple Select'),
-               ('text', 'Text'),
-               ('textarea', 'Text Area'),
-               ('yes_no', 'Checkbox'),
-               ('pills', 'Pills'),
-               ('select', 'Select'),
-               ('color', 'Color'),
-               ('multi', 'Multi-checkbox (option)'),
-          ],
-          help="The display type used in the Product Configurator.")
-
-     related_display_type = fields.Selection(
-          related='product_attribute_id.display_type',
-          store=True,
-          readonly=True
-     )
-
-     enable = fields.Boolean(string="Enable", default=True)
-     value_per_channel = fields.Boolean(string="Value per channel", default=False)
-     value_per_locale = fields.Boolean(string="Value per locale", default=False)
-
-     @api.depends('attr_group_id', 'product_attribute_id')
-     def _compute_used_attribute_ids(self):
-          """Compute all attributes already assigned to groups."""
-          all_used_attributes = self.env['attribute.group.lines'].search([]).mapped('product_attribute_id')
-          for line in self:
-               line.used_attribute_ids = all_used_attributes
-
-     @api.onchange('enable')
-     def _onchange_enable(self):
-          if self.product_attribute_id:
-               attribute_name = f"x_{self.product_attribute_id.name.lower().replace(' ', '_')}"
-               name = f"add_field_{attribute_name}_to_product_management"
-               attribute_view_exist = self.env['ir.ui.view'].search(
-                    [('name', '=ilike', name), ('active', '!=', None)])
-               arch = self._arch(attribute_name,self.enable)
-               if attribute_view_exist.arch:
-                    attribute_view_exist.arch = arch
-
-     def _arch(self,attribute_name,active):
-          if active == False:
-               new_field_xml = f'<field name="{attribute_name}" invisible="True"/>'
-          else:
-               new_field_xml = f'<field name="{attribute_name}"/>'
-          arch = f"""
-                       <xpath expr="//field[@name='{self.product_attribute_id.position_ref_field_id.name}']" position="after">
-                                                          %s
-                                                  </xpath>""" % new_field_xml
-          return arch
-
-
-class AttributeGroupUnlinkWizard(models.TransientModel):
-    _name = 'attribute.group.unlink.wizard'
-    _description = 'Wizard to Confirm Deletion of Attribute Group'
-
-    group_id = fields.Many2one('attribute.group', string="Attribute Group")
-
-    def confirm_unlink(self):
-        menu_id = self.env['ir.ui.menu'].search([('name', '=', 'Attributes Groups')])
-        if self.group_id:
-             self.group_id.unlink()
-        return {
-             'type': 'ir.actions.client',
-             'tag': 'reload',
-             'params': {
-                  'menu_id': menu_id.id,
-             },
-        }
-
 
 class ManufacturerAttribute(models.Model):
      _name = 'manufacturer.attribute'
@@ -901,29 +53,6 @@ class FamilyProducts(models.Model):
      attribute4_id = fields.Many2one('product.attribute4','',related='family_id.attribute4_id',)
      attribute4_val = fields.Char('Attribute 4')
      select_sku = fields.Boolean('Select')
-
-
-class ProductAttribute(models.Model):
-     _name = 'product.attribute1'
-
-     name = fields.Char('Name', required=True)
-
-
-class ProductAttribute2(models.Model):
-     _name = 'product.attribute2'
-
-     name = fields.Char('Name', required=True)
-
-
-class ProductAttribute3(models.Model):
-     _name = 'product.attribute3'
-
-     name = fields.Char('Name', required=True)
-     
-class ProductAttribute4(models.Model):
-     _name = 'product.attribute4'
-
-     name = fields.Char('Name', required=True)
 
 
 class FamilyAttribute(models.Model):
